+++ order_service/src/main/java/on/ssgdeal/order_service/application/dto/CancelOrderResponseDto.java
@@ -1,4 +1,4 @@
-package on.ssgdeal.order_service.application.service.dto;
+package on.ssgdeal.order_service.application.dto;
 
 public record CancelOrderResponseDto(Long orderId) {
