+++ order_service/src/main/java/on/ssgdeal/order_service/application/dto/CancelTotalOrderRequestDto.java
@@ -1,4 +1,4 @@
-package on.ssgdeal.order_service.application.service.dto;
+package on.ssgdeal.order_service.application.dto;
 
 public record CancelTotalOrderRequestDto(Long totalOrderId, LoginUserInfoDto loginUserInfo) {
