+++ order_service/src/main/java/on/ssgdeal/order_service/application/dto/GetTotalOrderDetailResponseDto.java
@@ -1,4 +1,4 @@
-package on.ssgdeal.order_service.application.service.dto;
+package on.ssgdeal.order_service.application.dto;
 
 import java.time.LocalDateTime;
 import java.util.Comparator;
