+++ order_service/src/main/java/on/ssgdeal/order_service/application/dto/TotalOrderProductInfo.java
@@ -1,4 +1,4 @@
-package on.ssgdeal.order_service.application.service.dto;
+package on.ssgdeal.order_service.application.dto;
 
 import java.util.List;
