+++ order_service/src/main/java/on/ssgdeal/order_service/domain/entity/dtos/mapper/TotalOrderEntityLayerMapper.java
@@ -1,8 +1,8 @@
 package on.ssgdeal.order_service.domain.entity.dtos.mapper;
 
-import on.ssgdeal.order_service.application.service.dto.LoginUserInfoDto;
-import on.ssgdeal.order_service.application.service.dto.UpdateCancelOrderSuccessRequestDto;
-import on.ssgdeal.order_service.application.service.dto.UpdateTotalOrderSuccessRequestDto;
+import on.ssgdeal.order_service.application.dto.LoginUserInfoDto;
+import on.ssgdeal.order_service.application.dto.UpdateCancelOrderSuccessRequestDto;
+import on.ssgdeal.order_service.application.dto.UpdateTotalOrderSuccessRequestDto;
 import on.ssgdeal.order_service.domain.entity.dtos.GetTotalOrdersUserInfoDto;
 import on.ssgdeal.order_service.domain.entity.dtos.UpdateCancelOrderSuccessDto;
 import on.ssgdeal.order_service.domain.entity.dtos.UpdateTotalOrderSuccessDto;
