+++ order_service/src/main/java/on/ssgdeal/order_service/presentation/internal/dto/UpdateTotalOrderFailRequest.java
@@ -1,7 +1,7 @@
 package on.ssgdeal.order_service.presentation.internal.dto;
 
 import jakarta.validation.constraints.NotNull;
-import on.ssgdeal.order_service.application.service.dto.UpdateTotalOrderFailRequestDto;
+import on.ssgdeal.order_service.application.dto.UpdateTotalOrderFailRequestDto;
 
 public record UpdateTotalOrderFailRequest(@NotNull Long totalOrderId) {
